# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegra",
    "description": "Production-ready Agent Protocol server",
    "version": "0.7.2"
  },
  "paths": {
...
    "/runs/wait": {
      "post": {
        "tags": [
          "Stateless Runs"
        ],
        "summary": "Stateless Wait For Run",
        "description": "Create a stateless run and wait for completion.\n\nGenerates an ephemeral thread, delegates to the threaded ``wait_for_run``\nendpoint, and deletes the thread on completion (unless\n``on_completion=\"keep\"``).",
        "operationId": "stateless_wait_for_run_runs_wait_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Stateless Wait For Run Runs Wait Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/stream": {
      "post": {
        "tags": [
          "Stateless Runs"
        ],
        "summary": "Stateless Stream Run",
        "description": "Create a stateless run and stream its execution.\n\nGenerates an ephemeral thread, delegates to the threaded\n``create_and_stream_run`` endpoint, and deletes the thread after the\nstream finishes (unless ``on_completion=\"keep\"``).",
        "operationId": "stateless_stream_run_runs_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events stream",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs": {
      "post": {
        "tags": [
          "Stateless Runs"
        ],
        "summary": "Stateless Create Run",
        "description": "Create a stateless background run.\n\nGenerates an ephemeral thread, delegates to the threaded ``create_run``\nendpoint, and schedules cleanup as a background task (unless\n``on_completion=\"keep\"``).",
        "operationId": "stateless_create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProtocolError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentProtocolError": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "AgentProtocolError",
        "description": "Standard Agent Protocol error response"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Run": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Unique identifier for the run."
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "Thread this run belongs to."
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id",
            "description": "Assistant that is executing this run."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current run status: pending, running, error, success, timeout, or interrupted.",
            "default": "pending"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "description": "Input data provided to the run."
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output",
            "description": "Final output produced by the run, or null if not yet complete."
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message",
            "description": "Error message if the run failed."
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Configuration passed to the graph at runtime.",
            "default": {}
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Context variables available during execution.",
            "default": {}
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Identifier of the user who owns this run."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when the run was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Timestamp when the run was last updated."
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "thread_id",
          "assistant_id",
          "input",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "Run",
        "description": "Run entity model\n\nStatus values: pending, running, error, success, timeout, interrupted"
      },
      "RunCreate": {
        "properties": {
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id",
            "description": "Assistant to execute"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input",
            "description": "Input data for the run. Optional when resuming from a checkpoint."
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config",
            "description": "Execution config",
            "default": {}
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Execution context",
            "default": {}
          },
          "checkpoint": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint",
            "description": "Checkpoint configuration (e.g., {'checkpoint_id': '...', 'checkpoint_ns': ''})"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Enable streaming response",
            "default": false
          },
          "stream_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Mode",
            "description": "Requested stream mode(s)"
          },
          "on_disconnect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "On Disconnect",
            "description": "Behavior on client disconnect: 'cancel' (default) or 'continue'."
          },
          "on_completion": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "delete",
                  "keep"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "On Completion",
            "description": "Behavior after stateless run completes: 'delete' (default) removes the ephemeral thread, 'keep' preserves it."
          },
          "multitask_strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multitask Strategy",
            "description": "Strategy for handling concurrent runs on same thread: 'reject', 'interrupt', 'rollback', or 'enqueue'."
          },
          "command": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command",
            "description": "Command for resuming interrupted runs with state updates or navigation"
          },
          "interrupt_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Before",
            "description": "Nodes to interrupt immediately before they get executed. Use '*' for all nodes."
          },
          "interrupt_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt After",
            "description": "Nodes to interrupt immediately after they get executed. Use '*' for all nodes."
          },
          "stream_subgraphs": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Subgraphs",
            "description": "Whether to include subgraph events in streaming. When True, includes events from all subgraphs. When False (default when None), excludes subgraph events. Defaults to False for backwards compatibility.",
            "default": false
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Request metadata (e.g., from_studio flag)"
          }
        },
        "type": "object",
        "required": [
          "assistant_id"
        ],
        "title": "RunCreate",
        "description": "Request model for creating runs"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}